+++ src/main/java/dev/jaxydog/astral/content/sound/CustomSoundEvents.java
@@ -1,9 +1,0 @@
-package dev.jaxydog.astral.content.sound;
-
-import dev.jaxydog.astral.register.ContentRegistrar;
-
-public final class CustomSoundEvents extends ContentRegistrar {
-
-	public static final CustomSoundEvent SPRAY_BOTTLE_USE = new CustomSoundEvent("item.spray_bottle.use");
-
-}